# internal/adapters/http/docs/swagger.yaml
basePath: /api
definitions:
...
  handler.BatchUpdateSettingsRequest:
    properties:
      settings:
        description: 至少需要一个设置项
        items:
          properties:
            key:
              type: string
            value:
              description: JSONB 原生值
          required:
          - key
          type: object
        minItems: 1
        type: array
    required:
    - settings
    type: object
  handler.CreateCategoryRequest:
    properties:
      icon:
        example: mdi-cog
        type: string
      key:
        example: custom
        type: string
      label:
        example: 自定义配置
        type: string
      order:
        example: 100
        type: integer
    required:
    - key
    - label
    type: object
  handler.CreateSettingRequest:
    properties:
      category_id:
        example: 1
        type: integer
      default_value: {}
      group:
        example: basic
        type: string
      key:
        example: site_name
        type: string
      label:
        example: 网站名称
        type: string
      order:
        example: 0
        type: integer
      ui_config:
        example: '{}'
        type: string
      value_type:
        example: string
        type: string
    required:
    - category_id
    - default_value
    - key
    type: object
  handler.SetUserSettingRequest:
    properties:
      value:
        description: JSONB 原生值
    type: object
  handler.UpdateCategoryRequest:
    properties:
      icon:
        example: mdi-settings
        type: string
      label:
        example: 更新后的标签
        type: string
      order:
        example: 50
        type: integer
    type: object
...
  handler.UpdateSettingRequest:
    properties:
      default_value: {}
      label:
        example: 更新后的标签
        type: string
      order:
        type: integer
      ui_config:
        type: string
    type: object
...
  response.DataResponse-array_setting_CategoryDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 响应数据
        items:
          $ref: '#/definitions/setting.CategoryDTO'
        type: array
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.DataResponse-array_setting_SettingsCategoryDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        description: 响应数据
        items:
          $ref: '#/definitions/setting.SettingsCategoryDTO'
        type: array
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.DataResponse-setting_CategoryDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/setting.CategoryDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-setting_SettingDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/setting.SettingDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
  response.DataResponse-setting_UserSettingDTO:
    properties:
      code:
        description: HTTP 状态码
        type: integer
      data:
        allOf:
        - $ref: '#/definitions/setting.UserSettingDTO'
        description: 响应数据
      error:
        description: 错误详情（仅失败时）
      message:
        description: 消息描述
        type: string
    type: object
...
  response.ErrorDetail:
    properties:
      code:
        description: 业务错误码（小写下划线）
        type: string
      details:
        description: 额外详情（如验证错误列表）
      message:
        description: 错误消息
        type: string
    type: object
  response.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorDetail'
    type: object
  response.MessageResponse:
    description: 纯消息响应格式
    properties:
      code:
        description: HTTP 状态码
        type: integer
      message:
        description: 消息描述
        type: string
    type: object
...
  setting.SettingsCategoryDTO:
    properties:
      category:
        type: string
      groups:
        items:
          $ref: '#/definitions/setting.SettingsGroupDTO'
        type: array
      icon:
        type: string
      label:
        type: string
    type: object
...
host: localhost:40012
info:
  contact:
    email: lwmacct@icloud.com
    name: API Support
    url: https://github.com/lwmacct/251117-go-ddd-template/issues
  description: 基于 DDD+CQRS 架构的 Go Web 应用 API 文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/lwmacct/251117-go-ddd-template
  title: Go DDD Template API
  version: 1.0.3
paths:
...
  /api/admin/settings:
    get:
      consumes:
      - application/json
      description: 获取按 Category → Group → Settings 层级组织的配置数据，用于前端动态渲染设置页面。支持按分类过滤（懒加载）。
      parameters:
      - description: 分类 Key（如 general），为空返回全量
        in: query
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 配置列表（层级结构）
          schema:
            $ref: '#/definitions/response.DataResponse-array_setting_SettingsCategoryDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 分类不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 配置列表
      tags:
      - Admin - Settings
    post:
      consumes:
      - application/json
      description: 管理员创建新的系统配置项
      parameters:
      - description: 配置信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateSettingRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 配置创建成功
          schema:
            $ref: '#/definitions/response.DataResponse-setting_SettingDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建配置
      tags:
      - Admin - Settings
  /api/admin/settings/{key}:
    delete:
      consumes:
      - application/json
      description: 管理员删除指定的系统配置项
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 配置删除成功
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 配置不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除配置
      tags:
      - Admin - Settings
    get:
      consumes:
      - application/json
      description: 根据配置键获取配置详情
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 配置详情
          schema:
            $ref: '#/definitions/response.DataResponse-setting_SettingDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 配置不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 配置详情
      tags:
      - Admin - Settings
    put:
      consumes:
      - application/json
      description: 管理员更新指定配置项的值和标签
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 配置更新成功
          schema:
            $ref: '#/definitions/response.DataResponse-setting_SettingDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 配置不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新配置
      tags:
      - Admin - Settings
  /api/admin/settings/batch:
    post:
      consumes:
      - application/json
      description: 管理员批量更新多个系统配置项的值
      parameters:
      - description: 配置列表
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.BatchUpdateSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 批量更新成功
          schema:
            $ref: '#/definitions/response.MessageResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 批量更新配置
      tags:
      - Admin - Settings
  /api/admin/settings/categories:
    get:
      consumes:
      - application/json
      description: 获取所有配置分类，按排序权重升序排列
      produces:
      - application/json
      responses:
        "200":
          description: 分类列表
          schema:
            $ref: '#/definitions/response.DataResponse-array_setting_CategoryDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 配置分类列表
      tags:
      - Admin - Setting Categories
    post:
      consumes:
      - application/json
      description: 管理员创建新的配置分类
      parameters:
      - description: 分类信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 分类创建成功
          schema:
            $ref: '#/definitions/response.DataResponse-setting_CategoryDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 创建配置分类
      tags:
      - Admin - Setting Categories
  /api/admin/settings/categories/{id}:
    delete:
      consumes:
      - application/json
      description: 管理员删除指定的配置分类（如有关联配置项则拒绝删除）
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: 分类删除成功
        "400":
          description: 存在关联配置项
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 分类不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除配置分类
      tags:
      - Admin - Setting Categories
    get:
      consumes:
      - application/json
      description: 根据 ID 获取配置分类详情
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分类详情
          schema:
            $ref: '#/definitions/response.DataResponse-setting_CategoryDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 分类不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 配置分类详情
      tags:
      - Admin - Setting Categories
    put:
      consumes:
      - application/json
      description: 管理员更新指定配置分类的信息（Key 不可修改）
      parameters:
      - description: 分类 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分类更新成功
          schema:
            $ref: '#/definitions/response.DataResponse-setting_CategoryDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 分类不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新配置分类
      tags:
      - Admin - Setting Categories
...
  /api/user/settings/{key}:
    delete:
      consumes:
      - application/json
      description: 删除用户自定义配置，恢复为系统默认值
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 重置成功
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 重置用户配置
      tags:
      - User - Settings
    get:
      consumes:
      - application/json
      description: 根据配置键获取用户配置（合并系统默认值）
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 配置详情
          schema:
            $ref: '#/definitions/response.DataResponse-setting_UserSettingDTO'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 配置不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 用户配置详情
      tags:
      - User - Settings
    put:
      consumes:
      - application/json
      description: 用户设置自定义配置值
      parameters:
      - description: 配置键
        in: path
        name: key
        required: true
        type: string
      - description: 配置值
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.SetUserSettingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功
          schema:
            $ref: '#/definitions/response.DataResponse-setting_UserSettingDTO'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: 配置不存在
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 设置用户配置
      tags:
      - User - Settings
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"